Support XMP:Rating (read/write)
What steps will reproduce the problem? (issue #206)

1. Rate the picture with Shotwell (or another application which
   writes XMP-xmp:Rating, but not the EXIF Rating/RatingPercent
   tags)
2. Import the rated picture into jBrout.
3. Inspect the rating of the picture in jBrout.

What is the expected output? What do you see instead?

The rating defined in Shotwell is expected to be available in
jBrout, whereas it is not.

In fact, Shotwell reads and updates only the XMP Rating tag,
whereas jBrout reads and updates only the EXIF
Rating/RatingPercent tags. (support for XMP write only was added
in the original patch for rating,  bug #95  released in r314, but
I suspect it was removed by  bug #129  in r335)

The logic of this patch is summarized hereafter:

- During jBrout Import/Refresh, when reading the rating from the
  file,
  + look at the XMP tag first, and update the EXIF tags if needed
  + then look at the EXIF tags, and update the XMP tag if needed
- During jBrout rating, when writing the rating to the file,
  + udpate the XMP tag and the EXIF tags

Please note that JBrout.conf["synchronizeXmp"] has not been used.

git-svn-id: https://jbrout.googlecode.com/svn/trunk@365 b4d1b9c4-0e25-0410-a282-091a75307ab2